chore: split high contrast to its own service
